flash messages

+++ app/src/Controller/ProductController.php
@@ -9,25 +9,36 @@ use Symfony\Component\Routing\Annotation\Route;
 
 use Doctrine\Persistence\ManagerRegistry;
 use App\Entity\Product;
+use App\Security\userDisabledException;
 
 class ProductController extends AbstractController {
 
+    public function main(): Response {
+        return $this->redirectToRoute('getProducts');
+    }
+
     public function getProducts(ManagerRegistry $doctrine): Response {
         $entityManager = $doctrine->getManager();
-        //$listProducts = $entityManager->getRepository(Product::class)->findBy(['enabled'=>1]);
         $listProducts = $entityManager->getRepository(Product::class)->findAll();
         return $this->render('products/products.html.twig', [
             'listProducts' => $listProducts,
         ]);
     }
 
-    public function showProduct($id, ManagerRegistry $doctrine): Response {
+    public function getSingleProduct($id, ManagerRegistry $doctrine): Response {
         $entityManager = $doctrine->getManager();
         $product = $entityManager->getRepository(Product::class)->find($id);
-
-        return $this->render('products/singleProduct.html.twig', [
-            'product' => $product,
-        ]);
+        if(!$product){
+            $this->addFlash('error', "The product doesn't exist");
+            return $this->redirectToRoute('getProducts');
+        } elseif(!$product->isEnabled()){
+            $this->addFlash('error', 'The product is not active, full information cannot be accessed');
+            return $this->redirectToRoute('getProducts');
+        } else {
+            return $this->render('products/singleProduct.html.twig', [
+                'product' => $product,
+            ]);
+        }
     }
 
     public function createProduct(Request $request, ManagerRegistry $doctrine): Response {
@@ -35,11 +46,10 @@ class ProductController extends AbstractController {
         $product = new \App\Entity\Product();
         $formProduct = $this->createForm(\App\Form\ProductType::class, $product);
         $formProduct->handleRequest($request);
-        // Filter decimals
         if($formProduct->isSubmitted() && $formProduct->isValid()){
-            $product->setEnabled(1);
             $entityManager->persist($product);
             $entityManager->flush();
+            $this->addFlash('success', 'The product had been created');
             return $this->redirectToRoute('getProducts');
         }
         return $this->render('products/productCreate.html.twig', [
@@ -50,25 +60,38 @@ class ProductController extends AbstractController {
     public function updateProduct($id, Request $request, ManagerRegistry $doctrine): Response {
         $entityManager = $doctrine->getManager();
         $product = $entityManager->getRepository(Product::class)->find($id);
-        $formProduct = $this->createForm(\App\Form\ProductType::class, $product);
-        $formProduct->handleRequest($request);
-        if($formProduct->isSubmitted() && $formProduct->isValid()){
-            $product->setEnabled(1);
-            $entityManager->persist($product);
-            $entityManager->flush();
+        if(!$product){
+            $this->addFlash('error', "The product doesn't exist");
             return $this->redirectToRoute('getProducts');
+        } elseif(!$product->isEnabled()){
+            $this->addFlash('error', 'The product is not active, cannot be modified');
+            return $this->redirectToRoute('getProducts');
+        } else {
+            $formProduct = $this->createForm(\App\Form\ProductType::class, $product);
+            $formProduct->handleRequest($request);
+            if($formProduct->isSubmitted() && $formProduct->isValid()){
+                $entityManager->flush();
+                $this->addFlash('success', 'The product had been updated');
+                return $this->redirectToRoute('getProducts');
+            }
+            return $this->render('products/productUpdate.html.twig', [
+                'formProduct' => $formProduct->createView(),
+            ]);
         }
-        return $this->render('products/productUpdate.html.twig', [
-            'formProduct' => $formProduct->createView(),
-        ]);
     }
 
     public function deleteProduct($id, ManagerRegistry $doctrine): Response {
         $entityManager = $doctrine->getManager();
         $product = $entityManager->getRepository(Product::class)->find($id);
-        $product->setEnabled(0);
-        $entityManager->remove($product);
-        $entityManager->flush();
-        return $this->redirectToRoute('getProducts');
+        if(!$product){
+            $this->addFlash('error', 'The product do not exist.');
+            return $this->redirectToRoute('getProducts');
+        } else {
+            $entityManager->remove($product);
+            $entityManager->flush();
+            $this->addFlash('success', 'The product had been deleted');
+            return $this->redirectToRoute('getProducts');
+        }
+
     }
 }
